excluding null conditions - combination works\!
