+++ test_autofit/unit/database/query/new_model/test_new_model.py
@@ -1,9 +1,4 @@
-import pytest
-
 from autofit.database import query_model as q
-
-
-# Need the model with and without child Named Queries
 
 
 class TestCombination:
